requirements.txt
# Core web framework and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
flask>=2.3.0
flask-cors>=4.0.0

# LangChain and AI dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langgraph>=0.1.0

# Vector store and embeddings
faiss-cpu>=1.7.4
numpy>=1.24.0

# Data processing and parsing
beautifulsoup4>=4.11.0
pydantic>=2.5.0

# Environment and configuration
python-dotenv>=1.0.0

# HTTP client for frontend
httpx>=0.25.0

# Frontend UI
gradio>=4.0.0

# Typing extensions
typing-extensions>=4.8.0